[PAASA-3590] windows support (alpha)
