--- utils/dup.go
@@ -1,9 +1,0 @@
-// +build !arm64
-
-package utils
-
-import "syscall"
-
-func Dup(from, to int) error {
-	return syscall.Dup2(from, to)
-}
